Fix VersionWithType Crossgen2 check for parameterized types
As Egor Bogatov discovered in

https://github.com/dotnet/runtime/pull/90441

Crossgen2 is unable to compile a method taking RuntimeTypeMethod[]
parameter because Crossgen2 decides that the array doesn't version
with the module being compiled (System.Private.CoreLib in this
case). This change modifies the relevant logic so that, for
parameterized types, we decide on versioning based on the
parameter type.

Thanks

Tomas