Infra fixes for Ivan's new JIT count test
This change combines Ivan's private branch 'jit-participants'
with a couple of infra fixes to make his test work. The fundamental
remaining piece is figuring out how to get rid of the commented out
line in XUnitWrapperGenerator.cs, @jkoritzinsky please advise how
to check the ProjectReference item for the ReferenceOutputAssembly
so that we ignore the reference in the merged wrapper generator
when it's set to false - i.o.w. when the project doesn't intend to
use the reference as a direct component of its managed executable,
it's rather intented as a payload that's manipulated in a special
manner at runtime.

I assume that, once we clarify this one remaining problem, I'll send
out the PR for the build script changes; that should subsequently
let Ivan adapt my changes to his test case based on his discretion
and make it fully functional both locally and in the lab.

Thanks

Tomas