Fix invalid PGO optimization path on Nuget cache containing spaces.
Since CMD does not strip quotes from arguments passed in with quotes
by default, if a Nuget package cache path containing a space is
passed to build-runtime.cmd, it will pass it into an additional CMake
argument with an additional set of quotes.

This causes string semantics to be disabled upon reaching the
second opening quote, thus separating the cache path into two
separate arguments incorrectly, resulting in the PGO data optimization
pointing to an invalid path.

This patch fixes this issue by simply removing the redundant quotes
when accepting the arguments in build-runtime.cmd.