Switch fx_ver_t::as_str from stringstream to append/to_string (#90652)
Stop using std::stringstream to convert fx_ver_t to a string. It pulls in std::locale (and other things) in all the host binaries when we really just need a simple conversion of the numbers to a string. In some binaries, it was also the only usage of stringstream. On my local Windows builds, this was a 15 to 40 kB decrease in size for host entry point components. In theory, this should also be faster, although this is not a particularly perf-sensitive function (most usage is for when tracing is enabled).