Add regression test for GitHub issue #74132
I have verified locally that the problem described in the issue
no longer exists in .NET 8 RC1, the bug was probably fixed earlier
this year by David Wrighton. I think the regression test described
in the issue is useful so I'm proposing to merge it in; this PR
will then close the issue.

Fixes: #74132

Thanks

Tomas