[wasm] Add blazor startup measurement to browser-bench
New blazor-frame project is slightly modified blazor template.

The browser bench measures the starup time of the blazor-frame app
inside the iframe. It tries to load it from the `blazor-frame/index.html`
relative url. So to enable the blazor measurements, build the blazor-frame
app separately and add a link to (or copy) the published wwwroot
to the browser-bench app bundle. This will be usually handled by our
measurements infrastructure.

Also fixed responses for missing files in the simple server.