JIT: Remove TYP_BOOL (#90981)
TYP_BOOL servces no purpose but to confuse today. We try to use it in some cases
to make assumptions about the value being 0/1, but this logic is not correct in
the face of unnormalized booleans. This change remove TYP_BOOL replacing its
uses by TYP_UBYTE. This fixes an issue where we forgot to handle TYP_BOOL.

Fix #90798