JIT: Fix various HW intrinsic lowerings for unused values (#91213)
In the test case we disable CSE but then hoisting hoists out a HWINTRINSIC node
that ends up not CSE'd. This results in a top level unused HWINTRINSIC node that
lowering didn't handle in a transformation. There's a bunch of transformations
that do not handle this correctly, so fix it in all of them.

I couldn't find a test case that does not require disabling CSE (and I didn't
include the existing one since it never finishes with the bug fixed), but I
wouldn't bet on that one doesn't exist, so I think we should backport this
anyway.

Fix #91173