Fix allocation in DateTime.TryParse when it returns false
On some failure paths, it was boxing a char or an int that would have been used as part of creating an exception on a Parse path but that's ignored on a TryParse path. This avoids that boxing by storing those values into a dedicated int field on the DateTimeResult. While doing this, I also consolidated several other fields, such that this not only avoids the boxing, it also results in a net reduction of the size of DateTimeResult (it's passed around by ref, so this only slightly reduces a zeroing burden).