Fix NativeAOT unhandled exception stack trace
The recent change to add a new exception handling mechanism to coreclr
has broken stack traces on unhandled exceptions in nativeaot.

This change fixes it by reverting the part of the change that caused the
problem and that turned out to not to be needed for the new exception
handling either.

Close #91298