jiterp: static_assert that atomic ops are (less likely) to go wrong
Not every C implementation guarantees that atomic operations on
arbtrary types are lock free.  So for example, casting between atomic_ushort* and
uint16_t* might not actually be ok to do.  We can't assert that
they're inter-castable, but at least assert that they're the same size
and that atomic_ushort is always lock-free. There might still be
restrictions (for example atomic_ushort might have to be aligned
differently) but this should at least catch obvious data corruption.