[Release/8.0] Fix x64 crossbuild on macOS arm64
The cross build is failing due to the cpufeatures.c being compiled in.
This file tries to extract cpu features using cpuid / getauxval that
don't make sense to execute in the cross tools.

This change disables compiling the cpufeatures.c for cross build and
changes the JitGetProcessorFeatures to return zero in this case.