Try removing imports from tests.mobile.targets
Mobile trimmingtests should already use the live illink with
the logic in project.csproj.template. The import of illink.targets
shouldn't be required, unless tests.mobile.targets are required
for other test scenarios.

This import was resulting in a double-import, and wasn't working
because liveILLink needs ToolsILLinkDir to be set, but it's not
set during the trimming test build.