Fix setting EH region on new loop header blocks
One loop canonicalization is disallow a loop `bottom` block from being
the `head` block of another loop. When this canonicalization is done,
a new `head` block is inserted above the loop `top` of the `head` block
loop, and it becomes the new `head` block. When doing so, the EH region
of `top` was being extended to include the new `head` block. This is
illegal if the `top` block is the beginning of a `try`, and there are
branches from outside the `try` to the block, as it will create an illegal
branch into the middle of a `try` region.

The solution is to re-use existing code in the creation of loop pre-headers
to set the proper EH region on the new `head` block.

So, the canonicalization does this (example):

```
BB10 // top of loop L00
BB11 -> BB10 // bottom of loop L00; head of loop L01
BB12 // top of loop L01
BB13 -> BB12 // bottom of loop L01
```

Here, we insert a new BB20 to split the dual purpose of BB11, as such:
```
BB10 // top of loop L00
BB11 -> BB10 // bottom of loop L00
BB20 // new head of loop L01
BB12 // top of loop L01
BB13 -> BB12 // bottom of loop L01
```

Note that if BB12 is a single-block `try` region, the new BB20 cannot have
the same `try` region index as BB12, since BB13 branches to BB12. So in this
case, BB20 needs to get the `try` index of the parent EH region of BB12.