Fix whole program optimization option for Windows
The recent change to cleanup the way some options in Windows builds of
the runtime are overriden missed the fact that the
CMAKE_INTERPROCEDURAL_OPTIMIZATION cannot contain generator expressions
and so it was always set to "off".

The fix is to set CMAKE_INTERPROCEDURAL_OPTIMIZATION_RELEASE and
CMAKE_INTERPROCEDURAL_OPTIMIZATION_RELWITHDEBUGINFO instead.

I have also missed two places before where we were still adding /GL-.