Remove uses of IlcTrimMetadata from the test tree
We want to be testing the shipping configuration.

I left it in two test projects because the uses there were too annoying to fix up. For System.Security, the problem is in non-trimmability of the APIs and would need RD.XML/whatever. For Vectors, the problem is in the use of `dynamic` in the tests due to laziness and I'm equally lazy to rewrite all of it.

Fixes #71506.