[mono][jit] Don't truncate the high bits when ldelema index is nint (#88986)
* [mono][interp] Remove dead code

* [tests] Move issue to interp only

* [mono][jit] Fix bounds check when index is nint

Previous code was assuming index is i4. Remove MONO_ARCH_EMIT_BOUNDS_CHECK on amd64 since it was doing a I4 comparison, while the index reg is i8. Use MONO_EMIT_DEFAULT_BOUNDS_CHECK instead also on amd64.

On llvm we might not be able to insert the sign extending because it apparently interferes with abcrem optimization. We therefore split OP_BOUNDS_CHECK into two separate opcodes, so, after abcrem, we know whether we have to insert the sext or not.

* [mono][interp] Add missing sign extend

Fix passing of i4 to wrapper accepting native int. The wrapper no longer does the sign extend in order to not lose high bits of native int.

* feedback