Allow using VS to run host tests when libraries use different config
When building the host tests from VS, there are dependencies on projects from the libraries subset. If the libraries are prebuilt with different configuration (typically `Release`) the build in VS breaks because it can't find the project's output.

This change modifies the import of generators (which are the problematic projects) to let them override the configuration of the referenced projects.

It then adds the ability to set environment based on the `-lc` parameter to the build script when starting VS.

Also updates the doc to describe how to use this.