SPMI: Handle "successful error cases" in near differ on arm64 (#91783)
After #89654 SPMI replay will succeed instead of result in replay errors
in expected error cases (such as BADCODE or EE exception). To support
diffing such contexts, we record zero-sized assembly that the near
differ uses. However, on arm64 there is some additional code that calls
repCompileMethod to make some additional adjustments to the code blob,
and in the "EE exception" cases we cannot replay this function,
resulting in crash during asmdiff. This fixes the problem by only making
the adjustments when we know there is any code.

An alternative solution could be to avoid invoking the neardiffer at all
in the succeeding error cases, but this seemed like an ok pragmatic
solution.

Fix #91257