Avoid marking property/event attributes multiple times (#92094)
This tracks whether properties/events have been marked yet, and
avoids marking them multiple times. Fixes a recursion issue where
marking a property or event can cause marking of attributes which
may cause marking of the same property/event.

Removes the `AlreadyMarked` assert because
https://github.com/dotnet/runtime/pull/84620 propagates the
`Kind` from methods to properties - so it can look as if a
property is being marked from a descriptor the first time it is
encountered.