Add trim analysis warnings for delegate creation
The ILLink analyzer was not producing warnings for annotated methods,
lambdas, or local functions accessed for delegate creation.

Some ExpectedWarnings for ILLink/NativeAot were updated to avoid
relying on the name of the compiler-generated methods for lambdas
or local functions, because the analyzer reports the warnings as:

IL2111: Method 'lambda expression' with parameters... is accessed
via reflection.