Fix register resolution handling of RefTypeUpperVectorRestore
After RefTypeUpperVectorRestore was handled, and an upper vector restore
IR node was added, LSRA incorrectly kept processing the
RefTypeUpperVectorRestore RefPosition. This led to incorrectly adding
a GTF_SPILL flag to the associated LCL_VAR. This in turn led some
LCL_VAR to be both last use and GTF_SPILL, which confused debug info processing:
it tried to "close" a range twice when it only expects to see a "close" operation
once.

Fixes #91215