Fix trimming of DebuggerDisplay with Name
The `Name` property of the `DebuggerDisplay` attribute accepts the
same format string as its `Value` property, but does not prevent
trimming members it references. Thanks to this fix, members referenced by
any of these two properties are not trimmed and can be displayed by a
debugger.

Additionally, this change adds references of three projects to the
illink solution file to fix its build errors.