Prevent restoring illink for native-binplace.proj (#92289)
Fixes https://github.com/dotnet/runtime/issues/92194. The
reference to illink from `native-binplace.proj`, built as a
reference of `build-native.proj`, was hitting a nuget bug with
static graph restore. The bug seems to be specific to something
about the project file (maybe the language-specific targets,
since `native-binplace.proj` imports the `Microsoft.NET.Sdk`, but
doesn't have a `csproj` extension).

Fixed by explicitly marking this as not a source project, which
will prevent the import of illink.targets.