Properly handle debugger-enumerating interior pointers and enregistered refs (#92313)
* ICorDebug:  Properly handle interior pointers

There is an ancient bug in the dac root walking code.  If we hit an interior pointer on a callstack that lives outside of the GC heap, we will report that pointer as-is to ICorDebug (specifically `CordbRefEnum::Next`).  This is despite the fact that ICorDebug *specifically* requests that the dac only enumerate pointers to real objects.  This non-gc pointer will be treated as a real object pointer and will cause a failed HRESULT in CordbRefEnum.  Since this call is wrapped with `IfFailThrow` it will halt all further processing of roots.

This code makes a small, targeted change to not enumerate non-GC pointers when the caller requests we specifically only enumerate pointers to real objects.

This is not a recent regression, but rather a bug that has likely existed since the original code was written.

* Fix issue with enregistered values

* Fix assert and enregistered ref reporting

* Fix issue from merge

---------

Co-authored-by: Juan Sebastian Hoyos Ayala <juan.hoyos@microsoft.com>
Co-authored-by: Juan Hoyos <19413848+hoyosjs@users.noreply.github.com>