Change GetParametersNoCopy to GetParametersAsSpan and use it in a few more places (#92310)
* Change GetParametersNoCopy to GetParametersAsSpan and use it  in a few more places

The internal GetParametersNoCopy avoids a defensive ParameterInfo[] copy, but it requires the callers to promise not to mutate the array. I'm changing the method to return a `ReadOnlySpan<ParameterInfo>`, which means a caller can't mutate it without using unsafe code, and enabling the compiler to flag any misuse.  I've then used it in a few more places. We might subsequently want to consider exposing it publicly.

* Update src/libraries/Common/src/System/Runtime/InteropServices/ComEventsMethod.cs