Fix arm64 fragment unwinding
A bug in the Windows arm64 unwinder that existed a long time ago has
caused problems with unwinding in functions split in multiple fragments
in case the location in the function was in a secondary fragment. At
that time, it was not discovered that it was a bug in the unwinder and
it got "fixed" in the runtime by always using the first fragment unwind
info. However, now it turned out that was actually incorrect in some
cases. Checking the current state of the Windows unwinder revealed that
a bug was fixed there that was causing the problem we were seeing.
Effectively ignoring all the shadow prolog unwind info in the secondary
fragments.
This change reverts the old fix after the unwinder was updated.