Simplify handling of properties in illink analyzer
Previously this was effectively visiting referenced properties twice for each reference - it would visit both the referenced get/set method, and the property itself.

This change only visits the get/set method, and handles the property warnings there by treating the getter/setter as annotated when the property is annotated.

This also allows some simplification of the override checking logic.