Add New Test to Track Number of Jitted Methods (#91235)
* New beginning for the Jitted Methods Counting Test.

* EXPERIMENTAL CHANGE ONLY: Added a command to print the whole directory
tree after running the tests, and adjusted the pipelines
configuration to only run that one.

* experimental change only: Discovered the jobs have dependencies on
other jobs. Disabled the runtime pipeline entirely, and only kept the
relevant outerloop platforms for my change to save resources.

* EXPERIMENT-ONLY CHANGE: Added the 'ls -R' directly to the Helix Post
Commands, and disabled the dev-innerloop pipeline to further speed up
this testing and debugging process.

* Maybe it doesn't like me removing the innerloop...?

* More experimenting... So we can't do just musl...

* Changed the test's methodology to use the JitInfo API, instead of
logging all the jitted methods into a file.

* Added the `AlwaysUseCrossgen2` property to the HelloWorld test app.

* Now printing the Jitted Methods to compare with my local machine.

* REVIEW-READY: Reenabled all the pipelines to see if there are any
other broken ones.

* Changed the Jitted methods counter to consider the whole process,
instead of just the current thread.

* DEBUG-ONLY CHANGE: We're getting more Jitted methods on Arm64 for some
reason, so making the test always fail to see the full CI log to begin
investigating what in the world is going on.

* Enabled musl outerloop pipelines because we were observing different
results there.

* Will you stop complaining now?

* READY: Reenabled the pipelines and disabled the test for R2R_CG2
branches on Arm64. It will require more investigation and we need the
other platforms to be able to run this test.

* Changed the jits-number limit.

* Changed the test to not require an external app anymore, and fixed
some small bad merges in the CI files when rebasing.

* DEBUG-ONLY CHANGE: Printing some logging to try to figure out why
Windows is causing problems, yet again...

* ...

* ...

* Due to how some machines are set up, it is actually possible for a
dotnet app to be fully prejitted, and therefore show 0 jitted methods,
so added that case to the test.