Use a different crossgen2 when running crossgen2 during our build than the crossgen2 that we are shipping
Instead of using the same crossgen2 project when we are building the product as we are shipping, repurpose the _crossarch copy of the build to be used in all scenarios. This avoids a race condition in our build where building the runtime and tool packs can fail because Microsoft.NETCore.App.Runtime is trying to R2R all assemblies while Microsoft.NETCore.App.Crossgen2 is re-publishing crossgen2 as a single-file self-contained app.